Add basic persistence logic and scryfall integration

+++ src/Application/Models/TribalCards.cs
@@ -4,8 +4,8 @@ namespace Application.Models {
     public class TribalCards {
         public string Tribe { get; set; }
         public int NumberOfCreatures { get; set; }
-        public double AvgRating { get; set; }
-        public double TribalRating { get; set; }
+        public decimal AvgRating { get; set; }
+        public decimal TribalRating { get; set; }
         public IEnumerable<string> Colors { get; set; }
     }
 }
